move the logic to compute cflags/ldflags outsite of NinjaWriter, so that we can use it from cffiwriter
